apps/backend_api/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "backend_api"
version = "0.1.0"
description = "FastAPI backend for the Handwritten LaTeX Converter"
readme = "README.md" # Optional: Add a README later if you want
requires-python = ">=3.10"
# List your core dependencies here - add others as needed from requirements.txt
# Ensure versions match your environment if necessary
dependencies = [
    "fastapi",
    "uvicorn[standard]", # Includes uvloop and httptools
    "sqlalchemy",
    "psycopg2-binary", # Or psycopg2 if you install C extensions separately
    "alembic",
    "celery[redis]",
    "python-dotenv",
    "boto3",
    "openai",
    # Add core_converter package if it's separate and needs to be installed
    # "core_converter @ file:///<path_to_core_converter_package>", 
]

[tool.setuptools.packages.find]
where = ["src"]  # Look for packages in the src directory 